requirements.txt
streamlit>=1.28.0
boto3>=1.34.0
pandas>=2.0.0
pyathena>=3.0.0
python-dotenv>=1.0.0
plotly>=5.0.0 